// src/types/Ranking.d.ts
export type BigFish = {
  ranking: number;
  memberId: string;
  userName: string;
  profileIcon: string;
  lureId: string;
  lureType: string;
  color: string;
  field: string;
  size: number;
};

export type FishingCount = {
  ranking: number;
  memberId: string;
  userName: string;
  profileIcon: string;
  fishingCount: number;
};